fixed size_t issues

+++ Curve/include/Curve.h
@@ -13,8 +13,8 @@ public:
     size_t getControlPointsNumber() const;
     size_t getCurvePointsNumber() const;
     void showControlPoints() const;
-    void binomialCoeffs(int n,std::vector<size_t> &_C);
-    void evaluateBezierCurve( std::vector<Vec3> &_curvePoints, int _nControlPoints, const std::vector<Vec3> &_controlPoints, const std::vector<size_t> &_C);
+    void binomialCoeffs(size_t n, std::vector<size_t> &_C);
+    void evaluateBezierCurve(std::vector<Vec3> &_curvePoints, size_t _nControlPoints, const std::vector<Vec3> &_controlPoints, const std::vector<size_t> &_C);
     //void renderGL() const;
 private:
     size_t m_numControlPoints=4;
